feat: filtering cleanup
